CMake: Add missing dependencies for tst_qmlbasicapp
The test executable needs to depend on the QML module and its plugin
because there is no direct link at the library level. The dependency
exists at the QML level, so we need to tell CMake or else the plugin
may not get built and that will result in failures due to not being
able to load the QML plugin at run time.

Pick-to: 6.2
Change-Id: Iafd1b8c15e3b2c227c7654d6fe39340e686f72a1
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>