Handle QScreen::physicalDotsPerInchChanged changes for the initial screen
We connect to the signal when the screens change (in
QQuickWindow::handleScreenChanged), but never for the initial screen.
Not connecting means that changing the DPR of the initial screen doesn't
re-render QQuickItems with the new DPR, which was in particular visible
for QQuickImage when displaying SVGs.

Task-number: QTBUG-94622
Change-Id: I44ff3f8f3713d5a7bba8b6b8b4d5ca14530fe373
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 7e39a2204382a78ac6ff1e32dfe29dcbce65004c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>