Update QQuickImage's DPR if re-loading the image didn't
An SVG doesn't have a DPR, but we need to reload and repaint it anyway
when the DPR of the screen changes. So store the DPR of the item explicitly
if reloading it didn't change the value (for instance, because sourceSize
was not set). Otherwise, DPR changes are only handled when moving from a
1.0 (the default) DPR display to a different display, but not when moving
back.

Fixes: QTBUG-94622
Change-Id: I9f0a847699ab027ef876e341b8c6a954a6167ab3
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 08838f434b55024d9d5a9252fc2a3fb782b087da)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>