Fix TARGET_PATH not being retrieved correctly for plugin on Android
The call to get_target_property() was malformed, which would have led to
the target path always being generated directly from the URI instead.
For all current uses of the command, that would have been correct
anyway, but fix the call anyway in case the situation changes in the
future.

The TARGET_PATH will become internal soon anyway, but we want the code
here to be correct before making that change in case it has to be
reverted later to make TARGET_PATH a public keyword in the API.

Task-number: QTBUG-95083
Pick-to: 6.2
Change-Id: I3a258e2f828295a155970f0ffab8a1504d6efa9c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>