Make TARGET_PATH an internal option for QML CMake commands
TARGET_PATH should always be computed from the URI. In the past, there
were some rare temporary cases where it was slightly different to
assist with transitioning some existing QML modules. No internal uses
of overriding the TARGET_PATH remain now though and user projects
should always use a TARGET_PATH based on the URI. Keep the
implementation in the public command in case an unexpected use case
for it arises, but make it clear that it is now an internal option.
Also remove it from qt_internal_add_qml_module() since we know
it is not needed there now.

Fixes: QTBUG-95083
Pick-to: 6.2
Change-Id: I75928fc194b5da64424ec8a838c1db4c0441edcf
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>