Don't create an animation timer when destroying the animation job
QQmlAnimationTimer::instance() takes an optional boolean "create"
parameter that defaults to true. Creating the timer when shutting down
the process requires access to thread-local storage, which might already
be destroyed as part of process shutdown, resulting in a segfault from
a Q_ASSERT.

We don't want to create any timer at this point.

Pick-to: 6.2
Change-Id: If73a6a3db138d01e818ee8cbbfa517a7ff14c59c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>