Document Application singleton
Move the documentation from Qt.application to the newly created
Application type entry. Add more examples to show how one can interact
with the singleton.
Now that we do not need to deal with documenting nested properties, we
can get rid of the table based format, and instead use proper
\qmlproperty and \qmlsignal commands.
The documentation of Qt.application now simply links to the Application
singleton documentation, with the additional note about properties being
undefined if QGuiApplication is not used.

Pick-to: 6.2
Task-number: QTBUG-95265
Change-Id: I55f8a1746d551430133a07680082282781880f45
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>