JS: Ensure that array keeps valid after length changes and fix concat
This is a partial revert of 6fa617524a6d0a2bc988e2dc70e8d719d1b9c282.
The reasoning there was wrong: Due to the ring structure of the array,
there might be further (non-undefined) elements at the start of the
allocated memory. Those need to be copied to. This patch therefore
reverts the change in 6fa617524a6d0a2bc988e2dc70e8d719d1b9c282 which
simply set the size, and restores the copying behavior again.

The actual fix for the crash in QTBUG-81037 requires a change to how we
set the array length: Previously, when the size increased,  we only
reinitialized the array (as a sparse array)  when the new size was
greater than a certain threshold. If the new size was smaller than that
threshold (but larger than the current alloc value), we would end up
with an inconsistent array: It was non-sparse, but had a smaller
capacity than size, leading to the memory corruption in concat when the
elements that should exist (but did not) were accessed. This patch
ensures that we now always resize the alloc buffer if necessary.

Task-number: QTBUG-81037
Fixes: QTBUG-90456
Change-Id: Ie193aa3d714121ce6e8203c4b663b9015715e025
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Maximilian Goldstein <max.goldstein@qt.io>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 64777c3ed9dc66e3510ce2139538b93ad68e9fbb)