Add device pixel ratio property to QQuickRenderTarget
Currently, in order to render a QtQuick scene in a HiDPI offscreen
texture without a render window, one need to fight Qt, for example by
returning a dummy window in QQuickRenderControl::renderWindowFor().

This change adds a device pixel ratio property to QQuickRenderTarget,
which can be used by the QQuickWindow to calculate the logical size of
the offscreen texture.

The main motivation behind this change is to allow KWin to render some
internal QtQuick scenes at HiDPI. Since kwin is the compositor, it does
not have any windows that can be returned from the renderWindowFor()
function.

Change-Id: Iefa7046385e750e15ca0e53b398b51a8e0aabad2
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>