Make CLASS_NAME default match the documentation for qt6_add_qml_plugin
The implementation did not obtain the class name from the backing
target, but it should have. Instead, it was considering an internal
target property on the plugin target itself, if it already existed.
That was a leftover from earlier work where the plugin target may have
been created directly by calling qt_add_plugin(), but it doesn't match
the behavior we document or support. Simplify the implementation and
make it match the documented behavior.

Pick-to: 6.2
Change-Id: Icd9ee1167e189715e8aff57f147734a2476f630c
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>