Sync shadereffect materialType cleanup between renderloops
The render loops behave differently: the threaded one cleans
the "material cache" (ill-named for legacy GL, it is correctly
called material _type_ cache on the newer RHI code path) every
time the window is unexposed. The basic one keeps it and only
cleans the cache (and so destroys the opaque QSGMaterialType
objects) only when the scenegraph (i.e. QSGNodes and related
geometry and materials) gets released.

The former is incorrect, because if the QSGMaterial does not go
away, the QSGMaterialType should stay valid too. Strictly speaking
this is not a fatal error because the QSGMaterialType pointers
are never dereferenced. However, potential reuse of the same
address due to a subsequent 'new QSGMaterialType' is an issue.

Just switch to the latter everywhere, and also, to be fully pedantic,
make the order uniform across all 3 render loops: first destroy the
scene tree (where each QSGGeometryNode destroys its QSGMaterial if
owned), then the cache (which owns the QSGMaterialTypes for the
now-gone shadereffect nodes).

Fixes: QTBUG-94844
Change-Id: Iebbf34d36b9eccca85e2557b045e326eef347beb
Reviewed-by: Andy Nichols <andy.nichols@qt.io>