Ensure qml and resource files are copied to build directory
Nothing depended on the qml and resource files that were copied to the
build directory. If the developer changed the source files without
re-running CMake, the copies in the build directory would be
out-of-date. Add a dedicated target for these copied files (needed only
for tooling) and make the backing target depend on it.

Fixes: QTBUG-95365
Change-Id: If32e45b25f9680a098103c9c05dcca9266030162
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 54e99531c8896da689b628309e1963683691d998)
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>