Fix plugin importing in tst_qmlbasicapp
We get no automatic static plugin importing and linking for QML, so we
have to do it manually ourselves. This includes all of the Qt QML
modules we reference too.

This also uncovered that the tst_qmlbasicapp test requires QtQuick, but
no check was added in ccd3907b12810742d8897bc51de6ddfaa6a5e619 where
the dependency was introduced. Add that check so that the test is now
only added if QtQuick is enabled.

And because the tst_qmlbasicapp test defines static plugins,
we need CMake 3.21 or later to avoid linker command line
ordering issues. Make adding the test conditional on that too.

Change-Id: I0f41352fb2fe9bd6548bf30eb8ed93366cefd263
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 40f4f8d6ece815f06d4c7784764d06284bddbe02)