Use QDateTime to get offsets for emscripten (wasm)
Since our wasm implementation doesn't have access to time-zone data by
default, but its QDateTime implementation does manage to get offsets
right for local time, use it as fall-back for V4's Date
implementation's DaylightSavingTA() and getLocalTZA().

This implementation might also be viable for other cases without
timezone support (and a way to reset it when the system zone has
changed), but we'll need to experiment to find out. For now, since we
have nothing better for wasm, use it there.

In passing, update a comment about a bug report against the ECMA spec
to say it's been fixed (and we're compatible with the result).

Fixes: QTBUG-95314
Pick-to: 6.2 5.15
Change-Id: I40c1537815ada950dc0b5cebd4d641f7bfc45bd9
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Lorn Potter <lorn.potter@gmail.com>