Fix static build broken by tooling target missing from export set
A QML module target doesn't need to link to its tooling target, it only
needs to depend on it to ensure that the tooling target is brought up
to date in the local build. The linking relationship currently added
for it is too strong and causes CMake to complain about the tooling
target being missing from the export set when doing a static build.
Replace it with a target dependency instead of creating a linking
relationship. This more accurately enforces the requirement without
requiring the tooling target to be installed.

Amends 54e99531c8896da689b628309e1963683691d998

Task-number: QTBUG-95365
Change-Id: I6a51f6022b617144e12ffb35f870d99fa3107d0a
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit c106f2693552a834cc53b219cb8cea7b97b49804)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>