Add NO_PLUGIN option to qt6_add_qml_module()
The new option allows the caller to state more clearly that they do not
want a plugin at all. Previously, the NO_CREATE_PLUGIN_TARGET option
had to be given to prevent the command from trying to create a plugin
target, but that option is meant for when the project will create the
plugin target later.

Now that the caller has a way to explicitly say they don't want a
plugin, require PLUGIN_TARGET to be given if NO_CREATE_PLUGIN_TARGET
is present. This forces the project to provide the name of the plugin
target that they will be responsible for creating.

Fixes: QTBUG-95141
Pick-to: 6.2
Change-Id: I84e036074f3e2785f17a8d33ad87ee0b7e016173
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>