macOS, GroupBox: ensure we read groupBoxPadding from a StyleItem
groupBoxPadding is a property on NativeStyle.GroupBox, which is
set as the background. So read the groupBoxPadding off the
backgroundItem rather than the control (unless the application
has set a custom background, hence the need to check that the
background is also a StyleItem.

This will silence a run-time warning about groupBoxPadding not being
found.

Fixes: QTBUG-94021
Cherry-pick: 6.2
Change-Id: I3ac8d308167c86f72394de89d3b862d98eaed0ff
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>