Prevent circular dependency with source files added as a resource
If a QML module target adds a .cpp source file as a resource too, that
source file will exist in both the source directory and the build
directory. The rule created to do that copy needs to depend on the
source file, but the logic CMake uses for the DEPENDS option in
add_custom_command() favors picking up a source file of the specified
name ahead of looking in the current source directory. That causes a
circular dependency. Pass an absolute path to DEPENDS instead so that
it always depends on the source directory's file in this scenario.

Pick-to: 6.2
Change-Id: I5b3ec238383831ff4658ec128a3d2b64734396a1
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>