Improve logic for finding base dir of QML modules in build directory
When a QML module has a library as its backing target, try to walk back
up the directory tree in the build directory to find the base
directory. The QML module's target path needs to match the path to its
qmldir below that base directory. This will allow other QML modules
that use the same base directory to be found by tooling like qmllint.
We warn if the directory structure doesn't conform to this requirement.

Fixes: QTBUG-95144
Pick-to: 6.2
Change-Id: I1842ae35fc220847fb1cc36402b286e107969541
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>