SelectionRectangle: ensure we track correctly which handle is being dragged
As it stood, we would check if the user dragged on the top-left handle
by checking which component was being used to instatiate the handle.
This goes wrong if both the top-left and the bottom-right handle
is instantiated from the same delegate component.

This patch will make it explicit which handle is being instatiated
instead.

Pick-to: 6.2
Fixes: QTBUG-95622
Change-Id: I592a99bf0e3daf0b566e07494aa2de69e010749d
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>