SelectionRectangle: don't accept new selection from press'n'hold on top of a handle
Don't start a new selection if the user does a press'n'hold on top
of a handle. We should allow this if the press is done on top
of the TableView. Otherwise the user will typically do a press'n'hold
on top of handle while preparing to drag it somewhere, only
to instead clear the current selection.

This patch will check the position of the press'n'hold before
starting a new selection.

Fixes: QTBUG-95629
Change-Id: Ic8711250075c85ee564ae115c60cff629710b570
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 20f2234136d34a0a44d17046daf8a1356361bcdf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>