SelectionRectangle: don't accept new selection from press'n'hold on top of a handle
Don't start a new selection if the user does a press'n'hold on top
of a handle. We should allow this if the press is done on top
of the TableView. Otherwise the user will typically do a press'n'hold
on top of handle while preparing to drag it somewhere, only
to instead clear the current selection.

This patch will check the position of the press'n'hold before
starting a new selection.

Pick-to: 6.2
Fixes: QTBUG-95629
Change-Id: Ic8711250075c85ee564ae115c60cff629710b570
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>