Fix linkage of defaultCodegenWarningInterface
We want one instance of the function's static variable, so the function
must not be static itself.
This still can lead to having multiple copies in the case where multiple
shared libraries are used. But this is a) not the case currently and b)
having one copy pare library is not problematic: The variable is
stateless, and only static to ease lifetime management of the pointer.

Change-Id: Iad3ea1ec74125446d926bc6a462866aeae936811
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>