Resolve enums before types when importing qmltypes
Type resolution is a straightforward single pass procedure. In certain
corner cases it couldn't find/resolve a type if that wasn't already
resolved. This happens to be the case when importing and resolving QtQml
types (such as QQmlEasingEnums::Type)

As we assume that this only happens with enumerations, we can resolve
them first and only after perform full type resolution for imported
scopes. If in fact there are other affected members, then a proper
second type resolution pass would be required instead

Pick-to: 6.2
Change-Id: Ic9606ed8bd058a1ae981085ac3e47a6f6a95ed7f
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>