Repurpose CustomCompileStep as NoBatching in QSGMaterial
Qt 5's CustomCompileStep has limited use in Qt 6: as there are no
virtual compile() type of functions to override in the API anymore,
its only effect now is disabling batching for nodes using materials
that have the flag set.

Coincidentally it turns out that some views in Creator's QML Profiler
rely on some quite advanced manual alteration of gl_Position.z in the
materials' vertex shader.

This is pretty bad if batching is involved, although in Qt 5 they
could get away by setting the internal undocumented qt_zRange
uniform to a magic value (1.0) to alter the behavior of the final
gl_Position override statement inserted by the shader rewriter
to batched materials' vertex shaders.

In Qt 6 this approach is not available, so the Profiler's rendering
becomes incorrect due to the vertex shader's own gl_Position.z
alterations being lost when batching is active. Instead, Creator 5
(that uses Qt 6) will set CustomCompileStep on its QSGMaterial
subclasses where relevant in order to never batch nodes with
these materials.

All in all this implies that there is a need for a better-named flag.
Therefore, add a NoBatching flag in place of CustomCompileStep, while
still keeping the latter as an alias until Qt 7.

Change-Id: Ib806fbfa55132fdd02765adece366f0175129ae8
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>