CMake: Set up extra dependencies for qml static plugins
If a project wants to avoid usage of qt_import_qml_plugins and instead
wants to link to the qtquick2plugin plugin target directly, it would
also have to explicitly link against the other plugin targets that are
its dependencies, which are qmlplugin, modelsplugin and
workerscriptsplugin.
This is needed to ensure the _init object libraries are linked to
as well.
Linking only to the static backing libraries is not sufficient,
because those don't have the _init libraries as dependencies, only
the plugins do.

Each qml plugin target should automatically bring its plugin link
dependencies via usage requirements.

To achieve that, when a static qml plugin is created, we now collect
all the private dependencies of the plugin backing lib or the plugin
itself (if it has no backing lib), check if each dependency is a qml
module, extract the module's associated plugin if it has one and
list it as a dependency of the new qml plugin.

Now a project can add a

 target_link_libraries(target PRIVATE Qt6::qtquick2plugin)

and that will implicitly call

 target_link_libraries(app PRIVATE Qt6::workerscriptplugin)
 target_link_libraries(app PRIVATE Qt6::modelsplugin)
 target_link_libraries(app PRIVATE Qt6::qmlplugin)

Note that the workerscript plugin is added manually as a conditional
dependency for the qml plugin if the workerscript feature is enabled.

To ensure that almost all dependencies are collected, we defer the
call when using CMake 3.19.

One implementation wart is that we need to ensure these new
dependencies don't get written to the plugin prl files, because qmake
has its own mechanism for generating Q_IMPORT_PLUGIN files. To do that,
we wrap the dependencies in a special genex that tells the build
system not to add them to the prl files.

Task-number: QTBUG-95609
Change-Id: Ia10109bb2286bd08b788ed3e54803ccfee18f369
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 6fd1216801f078f65be2cbc748cc459cb6912a4f)