macOS: ensure that you can set a custom button background
As it stood, the implicit size of a button would be empty
if assigning it a custom background with an implicit size.
This differs from the other styles.

This patch will follow the same logic as in e.g the Fusion
style, and use the maximum size of the background
and the content item as the buttons implicit size.

Task-number: QTBUG-95544
Change-Id: Ifaac9bea5814b2c7819906460019a85e951a063e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 20054ba885804e928726a5c515ab29306aff036b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>