Check for exposed instead of just visible in grabWindow
This issue is reproducable both in Qt 5 and 6. When calling grabWindow()
"early enough", such as from an Component.onCompleted handler, the
window is not ready for rendering (as it has not yet been exposed, there
is no scenegraph initialized, etc.). Checking the visible property is
not safe here, because it may well be true right from the start during
the window object's life, while telling nothing about the actual
low-level state. This becomes fatal in particular when one relies on
setting Window.visible to true from QML, instead of calling show() from
C++, because then the isVisible() check is clearly wrong (already true,
while the window is nowhere near to be ready for rendering) and leads to
hitting the wrong branch in the main condition of grabWindow().

It is probably safe to assume that the isVisible() check in grabWindow()
was never actually correct, and perhaps was an oversight. What it wants
to test for is being exposed (i.e. ready to render, with the scenegraph
and related machinery up and running), but it rather just tests the
visible property of the QWindow. It's just that in the majority of uses
that happens not to cause any problems in practice, either because
grabWindow() is only called once the window is up and running, or
because toggling visible is deferred to show() from C++ which masks the
underlying issue in the logic.

Switch over to isRenderable() which is a helper function based on the
exposed state.

Add an autotest case as well.

Pick-to: 6.2
Fixes: QTBUG-95393
Change-Id: Ia0193229f6b3980ff6bd51d85db1408017c43b38
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>