CMake: Fix building app qml modules on iOS with CMake and Xcode
A project with an executable and a qml module attached to it fails to
configure when targeting iOS with CMake and Xcode with the error:

CMake Error in CMakeLists.txt:
  The custom command generating
    qmltestapp_cmake/qqc_test_qmltyperegistrations.cpp
  is attached to multiple targets:
    qqc_test
    qqc_test_qmltyperegistration
  but none of these is a common dependency of the other(s).  This is not
  allowed by the Xcode "new build system".

When generating Xcode build files, CMake wants the generation of files
to be driven by one common target rather than 2 separate targets.

We haven't found a nice fix for this yet, so to circumvent the error
we make the ${target}_qmltyperegistration custom target depend on the
main ${target} when using the Xcode generator.

A consequence of this is that _qmllint targets will now build
${target} when invoked.

Pick-to: 6.2
Task-number: QTBUG-95763
Change-Id: I79623423b3dc0727fe5fed3a871176b04a56a6d5
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Craig Scott <craig.scott@qt.io>