Introduce generalized grouped properties
You can now use IDs to set up bindings on other objects than the one
currently being constructed. Such bindings are only accepted if the name
given as ID is a deferred property in the surrounding object. The best
way to achieve this is by giving ImmediatePropertyNames, as that will
make all non-immediate names deferred.

You are then expected to explicitly handle the resulting deferred
bindings, for example in componentComplete(). In order to simply apply
the bindings, you can call qmlExecuteDeferred().

Task-number: QTBUG-95117
Change-Id: Iedcf07543426f8f14c23cf53f6f3bcb186a342b0
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Maximilian Goldstein <max.goldstein@qt.io>
Reviewed-by: Andrei Golubev <andrei.golubev@qt.io>